64bit string length for scalar strings

+++ c/include/om_variable.h
@@ -132,15 +132,15 @@ OmString_t om_variable_get_string(const OmVariable_t* variable);
 
 /// Get the length of a scalar variable if written to a file.
 /// If the scalar is a string, we need to know the length of the string.
-size_t om_variable_write_scalar_size(uint16_t name_size, uint32_t children_count, OmDataType_t data_type, uint16_t string_length);
+size_t om_variable_write_scalar_size(uint16_t name_size, uint32_t children_count, OmDataType_t data_type, uint64_t string_length);
 
 /// Get the length of a string variable if written to a file.
-// size_t om_variable_write_string_size(uint16_t name_size, uint32_t children_count, uint16_t string_length);
+// size_t om_variable_write_string_size(uint16_t name_size, uint32_t children_count, uint64_t string_length);
 
 /// Write a scalar variable with name and children variables
-void om_variable_write_scalar(void* dst, uint16_t name_size, uint32_t children_count, const uint64_t* children_offsets, const uint64_t* children_sizes, const char* name, OmDataType_t data_type, const void* value, uint16_t string_length);
+void om_variable_write_scalar(void* dst, uint16_t name_size, uint32_t children_count, const uint64_t* children_offsets, const uint64_t* children_sizes, const char* name, OmDataType_t data_type, const void* value, uint64_t string_length);
 
-// void om_variable_write_string(void* dst, uint16_t name_size, uint32_t children_count, const uint64_t* children_offsets, const uint64_t* children_sizes, const char* name, const char* value, uint16_t string_length);
+// void om_variable_write_string(void* dst, uint16_t name_size, uint32_t children_count, const uint64_t* children_offsets, const uint64_t* children_sizes, const char* name, const char* value, uint64_t string_length);
 
 /// Get the size of meta attributes of a numeric array if written to a file. Does not contain any data. Only offsets for the actual data.
 size_t om_variable_write_numeric_array_size(uint16_t name_size, uint32_t children_count, uint64_t dimension_count);
